finished `build`

+++ src/index.ts
@@ -6,6 +6,8 @@ import * as errors from './errors.js';
 import glob from 'glob';
 import { z } from 'zod';
 import * as path from 'path';
+import { createWriteStream, existsSync } from 'fs';
+import archiver from 'archiver';
 
 const FilterArgumentDefinition = z.string().min(1).or(z.object({
     name: z.string(),
@@ -32,7 +34,7 @@ interface IFilters {
     [id: string]: Filter
 }
 
-async function evalFilters(config: zTypes.IConfigRequired, buildFlags: unknown) {
+async function evalFilters(config: zTypes.IConfigEvaluated, buildFlags: any) {
     const context = {
         buildFlags,
         config
@@ -100,19 +102,67 @@ async function evalFilters(config: zTypes.IConfigRequired, buildFlags: unknown) 
     return filters;
 }
 
-export async function build(config: zTypes.IConfigRequired, buildFlags: unknown) {
+export async function build(config: zTypes.IConfigEvaluated, buildFlags: any) {
+    const packPath = path.join(config.srcPath, 'packs');
+    const BPPath = path.join(packPath, 'BP');
+    const RPPath = path.join(packPath, 'RP');
+
+    if (config.packs.includes('BP') && !existsSync(BPPath)) throw errors.InternalError(errors.ErrorCode.InternalBPNotFound, `Missing behavior pack directory at \`${BPPath}\`. If you want to choose a different location, change the \`srcPath\` value in the config. If you do not want a behavior pack, remove \`BP\` from \`packs\` in the config.`);
+    if (config.packs.includes('RP') && !existsSync(RPPath)) throw errors.InternalError(errors.ErrorCode.InternalRPNotFound, `Missing resource pack directory at \`${RPPath}\`. If you want to choose a different location, change the \`srcPath\` value in the config. If you do not want a resource pack, remove \`RP\` from \`packs\` in the config.`);
+
     let filters: unknown;
 
     await fs.rm(config.outPath, { force: true, recursive: true });
 
     await Promise.all([
         evalFilters(config, buildFlags).then(res => filters = res),
-        fs.cp(path.join(path.join(config.srcPath, 'pack')), config.outPath, { recursive: true })
+        fs.cp(packPath, config.outPath, { recursive: true })
     ]);
 
     for (const filterToExecute of config.filters) {
         const filter = (filters as IFilters)[filterToExecute.id];
         if (!filter) throw errors.InternalError(errors.ErrorCode.InternalBuildFilterNotFound, `Cannot execute filter with id \`${filterToExecute.id}\` because it does not exist.`);
         filter.mainFunction();
+    }
+
+    if (!buildFlags.production && !buildFlags.package) {
+        // copy packs to com.mojang
+        if (config.packs.includes('BP')) (async () => {
+            const destPath = path.join(config.comMojangPath, 'development_behavior_packs', config.packName);
+            await fs.rm(destPath, { recursive: true, force: true });
+            await fs.cp(BPPath, destPath, { recursive: true });
+            console.log(`Copied \`${BPPath}\` to \`${destPath}\`. Use \`--production\` to disable this.`)
+        })();
+
+        if (config.packs.includes('RP')) (async () => {
+            const destPath = path.join(config.comMojangPath, 'development_resource_packs', config.packName);
+            await fs.rm(destPath, { recursive: true, force: true });
+            await fs.cp(RPPath, destPath, { recursive: true });
+            console.log(`Copied \`${RPPath}\` to \`${destPath}\`. Use \`--production\` to disable this.`)
+        })();
+    }
+
+    if (buildFlags.package) {
+        // zip dist to .mcaddon
+        if (existsSync('tmp')) throw errors.InternalError(errors.ErrorCode.InternalBuildPackageTmpFileExists, `Cannot package project. There is already a file at \`tmp\`.`);
+        const output = createWriteStream('tmp');
+        const archive = archiver('zip');
+
+        archive.on('warning', err => {
+            throw errors.InternalError(errors.ErrorCode.InternalBuildPackageArchiverWarning, err);
+        });
+
+        archive.on('error', err => {
+            throw errors.InternalError(errors.ErrorCode.InternalBuildPackageArchiverError, err);
+        });
+
+        archive.pipe(output);
+        archive.directory(config.outPath, false);
+
+        await archive.finalize();
+
+        const packagedArchivePath = path.join(config.outPath, `${config.packName}.mcaddon`);
+        await fs.rename('tmp', packagedArchivePath);
+        console.log(`Packaged files at \`${config.outPath}\` into archive \`${packagedArchivePath}\`.`);
     }
 }
